+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 ### Added
 - peer fields: description, address, asn
+- route fields: interface, source, time
 
 ### Fixed
 - handle date format for newer bird